requirements.txt
asgiref==3.7.2
backoff==2.2.1
branca==0.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
config==0.5.1
Django==5.0.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-cors-headers==4.3.0
django-extra-keywords==1.1.2
django-humanize==0.1.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-paypal==2.0
django-session-timeout==0.1.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
folium==0.15.0
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.2
opencage==2.3.1
packaging==23.2
phonenumbers==8.13.25
pillow==10.2.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pycountry==22.3.5
PyJWT==2.8.0
pyscopg2==66.0.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.6.0
